+++ modules/web-mercator/src/web-mercator-utils.d.ts
@@ -1,5 +1,8 @@
 // TODO - THE UTILITIES IN THIS FILE SHOULD BE IMPORTED FROM WEB-MERCATOR-VIEWPORT MODULE
 
+/** Mapbox default altitude */
+export const DEFAULT_ALTITUDE: number;
+
 /** Util functions **/
 export function zoomToScale(zoom: number): number;
 
@@ -93,13 +96,13 @@ type ProjectionParameters = {
 };
 
 /**
- * Valculates mapbox compatible projection matrix from parameters
-// Note: This matrix has variable fov (specified in radians)
+ * Calculates mapbox compatible projection matrix from parameters
  *
  * @param options.width Width of "viewport" or window
  * @param options.height Height of "viewport" or window
  * @param options.pitch Camera angle in degrees (0 is straight down)
- * @param options.altitude of camera in screen units
+ * @param options.fovy field of view in degrees
+ * @param options.altitude if provided, field of view is calculated using `altitudeToFovy()`
  * @param options.nearZMultiplier control z buffer
  * @param options.farZMultiplier control z buffer
  * @returns project parameters object
@@ -107,6 +110,7 @@ type ProjectionParameters = {
 export function getProjectionParameters(options: {
   width: number;
   height: number;
+  fovy?: number;
   altitude?: number;
   pitch?: number;
   nearZMultiplier?: number;
@@ -124,7 +128,8 @@ export function getProjectionParameters(options: {
  * @param options.width Width of "viewport" or window
  * @param options.height Height of "viewport" or window
  * @param options.pitch Camera angle in degrees (0 is straight down)
- * @param options.altitude of camera in screen units
+ * @param options.fovy field of view in degrees
+ * @param options.altitude if provided, field of view is calculated using `altitudeToFovy()`
  * @param options.nearZMultiplier control z buffer
  * @param options.farZMultiplier control z buffer
  * @returns 4x4 projection matrix
@@ -133,10 +138,31 @@ export function getProjectionMatrix(options: {
   width: number;
   height: number;
   pitch: number;
-  altitude: number;
+  fovy?: number;
+  altitude?: number;
   nearZMultiplier: number;
   farZMultiplier: number;
 }): number[];
+
+/**
+ *
+ * Convert an altitude to field of view such that the
+ * focal distance is equal to the altitude
+ *
+ * @param altitude - altitude of camera in screen units
+ * @return fovy field of view in degrees
+ */
+export function altitudeToFovy(altitude: number): number;
+
+/**
+ *
+ * Convert an field of view such that the
+ * focal distance is equal to the altitude
+ *
+ * @param fovy - field of view in degrees
+ * @return altitude altitude of camera in screen units
+ */
+export function fovyToAltitude(fovy: number): number;
 
 /**
  * Project flat coordinates to pixels on screen.